If we manually called channel.onopen, remove the handler.
Otherwise, we would end up calling the handler twice, in the case of firefox.  But not in the case of chrome.  Iuno.

Y'know, in the case of firefox, the first call of the onopen handler ended up having a few more functions available than the second call, from nodejs's perspective.  But the readystate didn't indicate anything other than "open".  This has not yet shown to be a real problem, but it might.