Remove stuff I pasted in that broke firefox/chromium compatibility.
DtlsSrtpKeyAgreement: false;

I don't remember why I pasted that in there, but it made it so that you couldn't create an offer in firefox and accept it in chromium.  The error was "Failed to set remote offer sdp: Called with SDP without SDES crypto."

Take out that constraint and we're all good.