Complete round trip works with fixed wrtc-mock.
wrtc-mock was declaring data channels open whenever we used setRemoteDescription, rather than waiting for the ICE candidates to get accepted.  This masked a bug in the kademlia code that caused the ICE candidates not to get delivered from Alice to Craig.  Fixing wrtc-mock caused the "complete round trip" test to fail.  Now that test succeeds, even with the fixed wrtc-mock.