Fix potential future scope issues.
I'm not entirely sure I understand the scope issue, to be honest.  To be safe, I'm assuming I have to make a function factory whenever we've got a situation where I'm making a callback that uses one of the arguments to the function.  Like, maybe that's the situation where the callback will use whatever the last value of the argument was, rather than the value that it was when the callback was created?  I swear, I've seen that happen!